requirements-colab.txt
# Requirements cho Google Colab (CUDA 12.0+)
# Training hoàn toàn trên Colab với GPU

# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Face Recognition
insightface>=0.7.3
facenet-pytorch>=2.5.2
mtcnn>=0.1.1

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0

# Data Processing
pandas>=2.0.0
pyyaml>=6.0

# Visualization & Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Search Engine (GPU version)
faiss-gpu>=1.7.4
